histogram

+++ gimage.h
@@ -78,7 +78,7 @@ class gImage
 		std::string Stats();
 		std::vector<long> Histogram();
 		//std::map<GIMAGE_CHANNEL,std::vector<unsigned> > Histogram(unsigned channels, unsigned scale);
-		std::vector<unsigned> Histogram(unsigned channel, unsigned &hmax);
+		std::vector<long> Histogram(unsigned channel, unsigned &hmax);
 
 		//Setters
 		void setInfo(std::string name, std::string value);
